refactor: move sixel related stuff into sixel.go
