+++ nav.go
@@ -18,15 +18,6 @@ import (
 	times "gopkg.in/djherbis/times.v1"
 )
 
-const (
-	gSixelBegin     = "\033P"
-	gSixelTerminate = "\033\\"
-)
-
-var (
-	gSixelFiller = '\u2800'
-)
-
 type linkState byte
 
 const (
