+++ ui.go
@@ -468,19 +468,6 @@ func (win *win) printDir(screen tcell.Screen, dir *dir, selections map[string]in
 	}
 }
 
-type sixel struct {
-	x, y, wPx, hPx int
-	str            string
-}
-
-type sixelScreen struct {
-	wpx, hpx     int
-	fontw, fonth int
-	sx           []sixel
-	lastFile     string
-	altFill      bool
-}
-
 type ui struct {
 	screen       tcell.Screen
 	sxScreen     sixelScreen
@@ -947,7 +934,7 @@ func (ui *ui) draw(nav *nav) {
 
 	ui.screen.Show()
 	if ui.menuBuf == nil && ui.cmdPrefix == "" && len(ui.sxScreen.sx) > 0 {
-		ui.showSixels()
+		ui.sxScreen.showSixels()
 	}
 
 }
@@ -1325,54 +1312,4 @@ func listMatchesMenu(ui *ui, matches []string) error {
 
 	ui.menuBuf = b
 	return nil
-}
-
-func (sxs *sixelScreen) clear() {
-	sxs.sx = nil
-}
-
-// fillers are used to control when tcell redraws the region where a sixel image is drawn.
-// alternating between bold("ESC [1m") and regular is to clear the image before drawing a new one.
-func (sxs *sixelScreen) filler(path string, l int) (fill string) {
-	if path != sxs.lastFile {
-		sxs.altFill = !sxs.altFill
-		sxs.lastFile = path
-	}
-
-	if sxs.altFill {
-		fill = "\033[1m"
-		defer func() {
-			fill += "\033[0m"
-		}()
-	}
-
-	fill += strings.Repeat(string(gSixelFiller), l)
-	return
-}
-
-func (sxs *sixelScreen) updateSizes(wc, hc int) {
-	var err error
-	sxs.wpx, sxs.hpx, err = getTermPixels()
-	if err != nil {
-		sxs.wpx, sxs.hpx = -1, -1
-		log.Printf("getting terminal pixel size: %s", err)
-	}
-
-	sxs.fontw = sxs.wpx / wc
-	sxs.fonth = sxs.hpx / hc
-}
-
-func (xsx *sixelScreen) pxToCells(x, y int) (int, int) {
-	return x/xsx.fontw + 1, y/xsx.fonth + 1
-}
-
-func (ui *ui) showSixels() {
-	var buf strings.Builder
-	buf.WriteString("\0337")
-	for _, sixel := range ui.sxScreen.sx {
-		buf.WriteString(fmt.Sprintf("\033[%d;%dH", sixel.y+1, sixel.x+1))
-		buf.WriteString(sixel.str)
-	}
-	buf.WriteString("\0338")
-	fmt.Print(buf.String())
 }
